+++ src/injected/visualization/svg-drawer.ts
