+++ src/tests/unit/tests/injected/visualization/drawer-provider.test.ts
@@ -13,7 +13,7 @@ import { DrawerUtils } from '../../../../../injected/visualization/drawer-utils'
 import { HighlightBoxDrawer } from '../../../../../injected/visualization/highlight-box-drawer';
 import { NullDrawer } from '../../../../../injected/visualization/null-drawer';
 import { SingleTargetDrawer } from '../../../../../injected/visualization/single-target-drawer';
-import { SVGDrawer } from '../../../../../injected/visualization/svg-drawer-v2';
+import { SVGDrawer } from '../../../../../injected/visualization/svg-drawer';
 
 describe('DrawerProviderTests', () => {
     let testObject: DrawerProvider;
